@@ -10,13 +10,13 @@
         }
     ],
     "require": {
-        "php": "^7.0",
+        "php": "^8.0",
         "ext-dom": "*",
         "ext-mbstring": "*",
         "ext-libxml": "*"
     },
     "require-dev": {
-        "phpunit/phpunit": "^6"
+        "phpunit/phpunit": "^9"
     },
     "autoload": {
         "psr-4": {"S25\\DomTrawler\\": "src/"}
